refactor(tx-pool): clarify authority ownership and commit flow
Separate tracked plans from live reads and the ordered Apply path. Centralize complete shard edits, cell detachment, capacity rejection, and relay reset accounting while preserving existing behavior.

Validation: strict workspace Clippy, 1431 Nextest tests passed with 2 skipped, workspace formatting, and documentation link checks.
