+++ tx-pool/docs/ARCHITECTURE.md
@@ -101,14 +101,17 @@ tie those responsibilities to independent behavioral checks.
 | Request protocol and completion | [controller](../src/service/controller.rs), [message](../src/service/message.rs), [dispatch](../src/service/dispatch.rs) |
 | Active jobs and direct requests | [Pool](../src/authority/service.rs), [execution](../src/authority/service/execution.rs), [submission](../src/authority/service/submission.rs) |
 | Policy preparation | [ingress](../src/authority/ingress.rs), [membership](../src/authority/membership.rs), [chain](../src/authority/chain.rs) |
-| Coupled mutation and resource ownership | [Store](../src/authority/store.rs), [budget](../src/authority/budget.rs) |
+| Live reads and prepared decisions | [Store](../src/authority/store.rs), [Plan and read set](../src/authority/store/plan.rs) |
+| Coupled mutation and resource ownership | [Apply](../src/authority/store/apply.rs), [budget](../src/authority/budget.rs) |
 | Retained representation and runnable work | [model](../src/authority/model.rs), [residency](../src/authority/residency.rs), [queues](../src/authority/queue.rs), [waiting](../src/authority/waiting.rs) |
 | Resolution and canonical checks | [jobs](../src/authority/jobs.rs), [verification](../src/verification.rs) |
 | Committed effects and relay reconstruction | [notice](../src/authority/notice.rs), [relay](../src/authority/relay.rs) |
 | Queries and mining output | [query](../src/authority/query.rs), [template](../src/authority/template.rs), [packing](../src/authority/packing.rs), [compiled causal graph](../src/authority/packing/graph.rs), [conditional ordering](../src/authority/packing/ordering.rs), [block assembler](../src/block_assembler/mod.rs) |
 | Startup, shutdown and disk format | [builder](../src/service/builder.rs), [persistence](../src/persisted.rs) |
 
-`Store` keeps observation validation and every coupled index write together.
+`Store` owns live reads and the shared guard primitives. Its Plan module owns
+original observations and intended changes; its Apply module keeps preflight,
+coupled writes, guard release and payload retirement in one ordered path.
 `Pool` coordinates retry, execution and lifecycle; its submission and execution
 files implement that same owner. Relay transport and cursor recovery belong to
 one relay module. Packing's child modules own numeric graph compilation and
