+++ tx-pool/src/authority/store.rs
@@ -1,19 +1,24 @@
-//! The only live owner/projection writer. Policy prepares a flat Plan; Apply
-//! prepares charges, locks, validates observations, reserves capacity, mutates
-//! and appends notices.
+//! Live owners, tracked reads and their shared synchronization primitives.
+//! [Plan] owns a prepared decision; [apply] keeps its preflight, coupled mutation,
+//! guard release and retirement together.
+mod apply;
+mod plan;
+
+use plan::same_weak;
+pub(super) use plan::{Edit, Plan, ReadSet};
+
 use super::{
-    budget::{Amount, Budget, Limits, OwnerDelta},
+    budget::{Amount, Budget, Limits},
     jobs::Job,
-    model::{DependencyKey, Entry, Error, FullReason, Phase, RelationKey, Status},
-    notice::{self, Batch, Class, Effect, Outbox},
+    model::{DependencyKey, Entry, Error, FullReason, Phase, RelationKey},
+    notice::Outbox,
     queue::{Queues, WorkStage},
 };
-use crate::{error::Reject, util::compact_packed};
+use crate::util::compact_packed;
 use ckb_app_config::TxPoolConfig;
 use ckb_network::PeerIndex;
 use ckb_snapshot::Snapshot;
 use ckb_types::{
-    core::BlockView,
     packed::{Byte32, OutPoint, ProposalShortId},
     prelude::*,
 };
@@ -44,327 +49,6 @@ pub(super) const WAIT: u8 = 4;
 pub(super) const CHILD: u8 = 8;
 const ACCEPTED_ROLES: u8 = INPUT | DEP | CHILD;
 const WAKE_PAGE: usize = 32;
-
-// Unique owner hashes in order, each paired with its old and new membership.
-type MemberChanges<T> = Vec<(Byte32, (T, T))>;
-
-/// Every decision read is kept, including negative owner/spender observations.
-/// Successful resolution keeps producer identities and point spender facts;
-/// membership separately observes complete reader and descendant relations.
-#[derive(Clone, Debug, Default)]
-pub(super) struct ReadSet {
-    owners: BTreeMap<Byte32, Option<Weak<Entry>>>,
-    spenders: BTreeMap<OutPoint, Option<Byte32>>,
-    relations: BTreeMap<RelationKey, Option<Weak<()>>>,
-    peers: BTreeMap<PeerIndex, Option<Weak<()>>>,
-    // Keep full-capture vectors out of ordinary sparse transaction read sets.
-    all: Option<Box<[u64; SHARDS]>>,
-    accepted: Option<Box<[u64; SHARDS]>>,
-}
-fn same_weak<T>(a: &Option<Weak<T>>, b: &Option<Weak<T>>) -> bool {
-    match (a, b) {
-        (None, None) => true,
-        (Some(a), Some(b)) => a.ptr_eq(b),
-        _ => false,
-    }
-}
-
-/// Preserve the first observation of each key, rejecting a different reread.
-fn merge_observations<K: Ord + Clone, V: Clone>(
-    own: &mut BTreeMap<K, V>,
-    incoming: &BTreeMap<K, V>,
-    same: impl Fn(&V, &V) -> bool,
-) -> Result<(), Error> {
-    for (key, value) in incoming {
-        if let Some(original) = own.get(key) {
-            if !same(original, value) {
-                return Err(Error::Stale);
-            }
-        } else {
-            own.insert(key.clone(), value.clone());
-        }
-    }
-    Ok(())
-}
-
-impl ReadSet {
-    fn observed_owner(&self, hash: &Byte32) -> Option<&Option<Weak<Entry>>> {
-        self.owners.get(hash)
-    }
-    pub(super) fn observe_owner(
-        &mut self,
-        hash: &Byte32,
-        entry: Option<&Arc<Entry>>,
-    ) -> Result<(), Error> {
-        let observed = entry.map(Arc::downgrade);
-        if let Some(old) = self.owners.get(hash) {
-            if !same_weak(old, &observed) {
-                return Err(Error::Stale);
-            }
-        } else {
-            self.owners.insert(compact_packed(hash), observed);
-        }
-        Ok(())
-    }
-    pub(super) fn merge(&mut self, other: &Self) -> Result<(), Error> {
-        merge_observations(&mut self.owners, &other.owners, same_weak)?;
-        merge_observations(&mut self.spenders, &other.spenders, PartialEq::eq)?;
-        merge_observations(&mut self.relations, &other.relations, same_weak)?;
-        merge_observations(&mut self.peers, &other.peers, same_weak)?;
-        for (own, incoming) in [
-            (&mut self.all, &other.all),
-            (&mut self.accepted, &other.accepted),
-        ] {
-            if let Some(incoming) = incoming {
-                if own.as_ref().is_some_and(|old| old != incoming) {
-                    return Err(Error::Stale);
-                }
-                if own.is_none() {
-                    *own = Some(incoming.clone());
-                }
-            }
-        }
-        Ok(())
-    }
-    pub(super) fn into_verification_reads(self) -> Self {
-        Self {
-            owners: self.owners,
-            spenders: self.spenders,
-            ..Self::default()
-        }
-    }
-    pub(super) fn spent(&self) -> impl Iterator<Item = (&OutPoint, &Byte32)> {
-        self.spenders
-            .iter()
-            .filter_map(|(point, spender)| spender.as_ref().map(|hash| (point, hash)))
-    }
-}
-
-#[derive(Clone)]
-pub(super) struct Edit {
-    pub(super) before: Option<Arc<Entry>>,
-    pub(super) after: Option<Arc<Entry>>,
-}
-/// One decision owns its original observations, owner changes and obligations.
-/// Producers can add observations, but cannot replace them or edit the outcome
-/// vectors directly. A lifecycle revision stays paired with its snapshot.
-#[derive(Clone)]
-pub(super) struct Plan {
-    view: u64,
-    reads: ReadSet,
-    edits: BTreeMap<Byte32, Edit>,
-    effects: Vec<Effect>,
-    class: Class,
-    snapshot: Option<Arc<Snapshot>>,
-    dry_run: bool,
-    clear_all: bool,
-    committed: Vec<(ProposalShortId, Byte32)>,
-    ban: Option<(PeerIndex, Instant)>,
-    peer_access: Option<(PeerIndex, bool)>,
-    pub(super) wake: BTreeSet<DependencyKey>,
-    wake_advance: Option<WakePage>,
-}
-impl Plan {
-    pub(super) fn new(view: u64, class: Class, reads: ReadSet) -> Self {
-        Self {
-            view,
-            reads,
-            edits: BTreeMap::new(),
-            effects: Vec::new(),
-            class,
-            snapshot: None,
-            dry_run: false,
-            clear_all: false,
-            committed: Vec::new(),
-            ban: None,
-            peer_access: None,
-            wake: BTreeSet::new(),
-            wake_advance: None,
-        }
-    }
-    pub(super) fn edit(
-        &mut self,
-        before: Option<Arc<Entry>>,
-        after: Option<Arc<Entry>>,
-        effect: Option<Effect>,
-    ) -> Result<(), Error> {
-        let hash = before
-            .as_ref()
-            .or(after.as_ref())
-            .ok_or(Error::Fault("empty owner edit"))?
-            .hash();
-        if after.as_ref().is_some_and(|entry| entry.hash() != hash) {
-            return Err(Error::Fault("owner hash change"));
-        }
-        if self.edits.contains_key(&hash) {
-            return Err(Error::Fault("duplicate owner edit"));
-        }
-        self.reads.observe_owner(&hash, before.as_ref())?;
-        // An identical immutable owner is only a read. Re-inserting its queue
-        // projection could duplicate work that a worker has already selected.
-        if before
-            .as_ref()
-            .zip(after.as_ref())
-            .is_some_and(|(before, after)| Arc::ptr_eq(before, after))
-        {
-            self.effects.extend(effect);
-            return Ok(());
-        }
-        self.edits
-            .insert(compact_packed(&hash), Edit { before, after });
-        self.effects.extend(effect);
-        Ok(())
-    }
-    /// Notice-only outcomes have no owner edit, but still validate this Plan.
-    pub(super) fn notify(&mut self, effect: Effect) {
-        self.effects.push(effect);
-    }
-    pub(super) fn peer_banned(&mut self, store: &Store, peer: PeerIndex) -> Result<bool, Error> {
-        let banned = store.peer_banned(peer);
-        if self
-            .peer_access
-            .is_some_and(|original| original != (peer, banned))
-        {
-            return Err(Error::Stale);
-        }
-        self.peer_access = Some((peer, banned));
-        Ok(banned)
-    }
-    pub(super) fn ban_peer(
-        &mut self,
-        hash: &Byte32,
-        reject: Reject,
-        peer: PeerIndex,
-        deadline: Instant,
-    ) -> Result<(), Error> {
-        if self.ban.is_some() {
-            return Err(Error::Fault("duplicate peer ban"));
-        }
-        let effect = Effect::banned(hash, reject, peer, deadline)?;
-        self.ban = Some((peer, deadline));
-        self.notify(effect);
-        Ok(())
-    }
-    /// Resetting the lifecycle and invalidating relay knowledge are one outcome.
-    pub(super) fn reset(&mut self, snapshot: Arc<Snapshot>, clear_all: bool) {
-        self.snapshot = Some(snapshot);
-        self.clear_all = clear_all;
-        self.notify(Effect::reset());
-    }
-    /// Block observations and ordered committed-hash records come from the same
-    /// attached blocks; source preparation cannot append one without the other.
-    pub(super) fn chain(
-        &mut self,
-        snapshot: Arc<Snapshot>,
-        blocks: impl IntoIterator<Item = Arc<BlockView>>,
-    ) {
-        self.snapshot = Some(snapshot);
-        let blocks: Vec<_> = blocks.into_iter().collect();
-        for block in &blocks {
-            for transaction in block.transactions().iter().skip(1) {
-                self.committed.push((
-                    compact_packed(&transaction.proposal_short_id()),
-                    compact_packed(&transaction.hash()),
-                ));
-            }
-        }
-        if !blocks.is_empty() {
-            self.notify(Effect::blocks(blocks));
-        }
-    }
-    #[cfg(test)]
-    pub(super) fn committed_for_test(&mut self, records: Vec<(ProposalShortId, Byte32)>) {
-        self.committed = records;
-    }
-    pub(super) fn get(
-        &mut self,
-        store: &Store,
-        hash: &Byte32,
-    ) -> Result<Option<Arc<Entry>>, Error> {
-        store.get(hash, &mut self.reads)
-    }
-    /// Graph policy reuses the first immutable observation, including absence.
-    pub(super) fn original(
-        &mut self,
-        store: &Store,
-        hash: &Byte32,
-    ) -> Result<Option<Arc<Entry>>, Error> {
-        match self.reads.observed_owner(hash) {
-            Some(None) => Ok(None),
-            Some(Some(owner)) => owner.upgrade().map(Some).ok_or(Error::Stale),
-            None => self.get(store, hash),
-        }
-    }
-    pub(super) fn observe_owner(
-        &mut self,
-        hash: &Byte32,
-        entry: Option<&Arc<Entry>>,
-    ) -> Result<(), Error> {
-        self.reads.observe_owner(hash, entry)
-    }
-    pub(super) fn spender(
-        &mut self,
-        store: &Store,
-        point: &OutPoint,
-    ) -> Result<Option<Byte32>, Error> {
-        store.spender(point, &mut self.reads)
-    }
-    pub(super) fn members(
-        &mut self,
-        store: &Store,
-        key: &RelationKey,
-        roles: u8,
-    ) -> Result<Vec<Byte32>, Error> {
-        store.members(key, roles, &mut self.reads)
-    }
-    pub(super) fn peer_members(
-        &mut self,
-        store: &Store,
-        peer: PeerIndex,
-    ) -> Result<Vec<Byte32>, Error> {
-        store.peer_members(peer, &mut self.reads)
-    }
-    pub(super) fn capture_accepted(&mut self, store: &Store) -> Result<Vec<Arc<Entry>>, Error> {
-        let (view, _, entries, reads) = store.capture(true);
-        if view != self.view {
-            return Err(Error::Stale);
-        }
-        self.reads.merge(&reads)?;
-        for entry in &entries {
-            self.observe_owner(&entry.hash(), Some(entry))?;
-        }
-        Ok(entries)
-    }
-    /// A refused policy keeps every premise, including peer eligibility, while
-    /// dropping speculative owners and notices before constructing its rejection.
-    pub(super) fn discard_changes(self) -> Self {
-        Self {
-            peer_access: self.peer_access,
-            dry_run: self.dry_run,
-            ..Self::new(self.view, self.class, self.reads)
-        }
-    }
-    pub(super) fn dry_run(mut self) -> Self {
-        self.dry_run = true;
-        self.effects.clear();
-        self
-    }
-    #[cfg(any(test, feature = "internal"))]
-    pub(super) fn edits(&self) -> &BTreeMap<Byte32, Edit> {
-        &self.edits
-    }
-    #[cfg(test)]
-    pub(super) fn effects(&self) -> &[Effect] {
-        &self.effects
-    }
-    #[cfg(any(test, feature = "internal"))]
-    pub(super) fn silence_fixture(&mut self) {
-        self.effects.clear();
-    }
-    pub(super) fn advance(&mut self, page: WakePage) {
-        self.wake_advance = Some(page);
-    }
-}
 
 struct View {
     snapshot: Arc<Snapshot>,
@@ -505,12 +189,6 @@ impl<T> Guard<'_, T> {
         }
     }
 }
-/// Keep recoverable Result/Option propagation out of the post-preflight tail.
-/// This unit-return boundary does not claim panic or allocation-failure recovery.
-fn commit_infallibly(commit: impl FnOnce()) {
-    commit();
-}
-
 /// Temporary lock footprint: present bits select shards; write bits only upgrade.
 #[derive(Default)]
 struct LockFootprint {
@@ -599,61 +277,6 @@ fn compact_relation(key: &RelationKey) -> RelationKey {
         RelationKey::Dependency(key) => RelationKey::Dependency(compact_dependency(key)),
         RelationKey::Children(hash) => RelationKey::Children(compact_packed(hash)),
     }
-}
-fn visit_roles(entry: &Entry, mut add: impl FnMut(RelationKey, u8)) {
-    match &entry.phase {
-        Phase::Accepted(accepted) => {
-            for point in entry.transaction.input_pts_iter() {
-                add(RelationKey::Dependency(DependencyKey::Cell(point)), INPUT);
-            }
-            for point in accepted.dependencies() {
-                add(RelationKey::Dependency(DependencyKey::Cell(point)), DEP);
-            }
-            for hash in &accepted.parents {
-                add(RelationKey::Children(hash.clone()), CHILD);
-            }
-        }
-        Phase::Waiting(keys) | Phase::Replaced { triggers: keys, .. } => {
-            for key in keys {
-                add(RelationKey::Dependency(key.clone()), WAIT);
-            }
-        }
-        Phase::Resolve | Phase::Verify(_) => {}
-    }
-}
-// A roleless side cannot cancel a visited role. Stream that common transition
-// directly; two role-bearing sides still need a bounded owner-local merge.
-fn visit_role_changes(edit: &Edit, mut add: impl FnMut(RelationKey, u8, u8)) {
-    let has_roles = |entry: &&Entry| !matches!(entry.phase, Phase::Resolve | Phase::Verify(_));
-    let before = edit.before.as_deref().filter(has_roles);
-    let after = edit.after.as_deref().filter(has_roles);
-    match (before, after) {
-        (None, Some(after)) => visit_roles(after, |key, role| add(key, 0, role)),
-        (Some(before), None) => visit_roles(before, |key, role| add(key, role, 0)),
-        (Some(before), Some(after)) => {
-            let mut roles: BTreeMap<RelationKey, (u8, u8)> = BTreeMap::new();
-            visit_roles(before, |key, role| roles.entry(key).or_default().0 |= role);
-            visit_roles(after, |key, role| roles.entry(key).or_default().1 |= role);
-            for (key, (old, new)) in roles {
-                if old != new {
-                    add(key, old, new);
-                }
-            }
-        }
-        (None, None) => {}
-    }
-}
-fn peer(entry: &Entry) -> Option<PeerIndex> {
-    entry
-        .preaccepted()
-        .then(|| entry.source.residency_peer())
-        .flatten()
-}
-fn deadline(entry: &Entry) -> Option<Instant> {
-    entry
-        .preaccepted()
-        .then(|| entry.source.deadline())
-        .flatten()
 }
 
 impl Store {
@@ -1303,625 +926,6 @@ impl Store {
         result
     }
 
-    pub(super) fn apply(&self, mut plan: Plan) -> Result<Option<Arc<Batch>>, Error> {
-        let mut notice = None;
-        self.apply_plan(&mut plan, &mut notice)
-    }
-
-    /// Optional replacement history may lose capacity to another admission.
-    /// Keep this plan's exact notices for one synchronous retry; no reservation
-    /// or failed plan escapes to the service's asynchronous capacity wait.
-    pub(super) fn apply_admission(
-        &self,
-        mut plan: Plan,
-        retain_history: &mut bool,
-    ) -> Result<Option<Arc<Batch>>, Error> {
-        let mut notice = None;
-        match self.apply_plan(&mut plan, &mut notice) {
-            Err(Error::Full(FullReason::Pipeline | FullReason::History)) if *retain_history => {
-                *retain_history = false;
-                if self.is_stopped() {
-                    // Return to the caller's existing close/requeue handling.
-                    return Err(Error::Stale);
-                }
-                for edit in plan.edits.values_mut() {
-                    if edit
-                        .before
-                        .as_ref()
-                        .is_some_and(|old| old.accepted().is_some())
-                        && edit
-                            .after
-                            .as_ref()
-                            .is_some_and(|new| matches!(new.phase, Phase::Replaced { .. }))
-                    {
-                        edit.after = None;
-                    }
-                }
-                self.apply_plan(&mut plan, &mut notice)
-            }
-            result => result,
-        }
-    }
-
-    /// The private caller keeps a notice paired with its original plan and
-    /// never changes the effects or class between attempts. No VM, provider,
-    /// await, endpoint or payload destruction under the authority cut; all
-    /// expected failures precede the first owner/index mutation.
-    #[expect(
-        clippy::arithmetic_side_effects,
-        reason = "Revision deltas are prechecked; exact validated owner edits preserve bounded phase counts."
-    )]
-    fn apply_plan(
-        &self,
-        plan: &mut Plan,
-        notice: &mut Option<notice::Reservation>,
-    ) -> Result<Option<Arc<Batch>>, Error> {
-        #[cfg(feature = "profiling")]
-        let _span = tracing::trace_span!(target: "ckb_tx_pool_profile", "tx_pool.authority.apply")
-            .entered();
-        if self.is_faulted() {
-            return Err(Error::Fault("closed generation"));
-        }
-        let owner_delta = OwnerDelta::new(
-            plan.edits
-                .values()
-                .filter_map(|edit| edit.before.as_deref()),
-            plan.edits.values().filter_map(|edit| edit.after.as_deref()),
-        )?;
-        if notice.is_none() && !plan.dry_run {
-            *notice = self
-                .outbox
-                .reserve(std::mem::take(&mut plan.effects), plan.class)?;
-        }
-        let mut owner_footprint = LockFootprint::default();
-        let mut dependency_footprint = LockFootprint::default();
-        let mut peer_footprint = LockFootprint::default();
-        let mut owner_edits = Vec::with_capacity(plan.edits.len());
-        // Plan hashes are unique and ordered; the last per-key change can merge
-        // repeated roles of the current owner without another lookup tree.
-        let mut delta: BTreeMap<RelationKey, MemberChanges<u8>> = BTreeMap::new();
-        let mut peer_delta: BTreeMap<PeerIndex, MemberChanges<bool>> = BTreeMap::new();
-        for (hash, edit) in &plan.edits {
-            let index = self.owner_shard(hash);
-            owner_footprint.insert(index, true);
-            owner_edits.push((index, hash, edit));
-            if plan.clear_all {
-                continue;
-            }
-            visit_role_changes(edit, |key, old, new| {
-                if let Some(changes) = delta.get_mut(&key) {
-                    // Ordered owners keep duplicate roles in the final item.
-                    if let Some((_, roles)) = changes.last_mut().filter(|(last, _)| last == hash) {
-                        roles.0 |= old;
-                        roles.1 |= new;
-                    } else {
-                        changes.push((hash.clone(), (old, new)));
-                    }
-                } else {
-                    delta.insert(compact_relation(&key), vec![(hash.clone(), (old, new))]);
-                }
-                dependency_footprint.insert(self.route(&key), (old | new) & INPUT != 0);
-            });
-            let old_peer = edit.before.as_deref().and_then(peer);
-            let new_peer = edit.after.as_deref().and_then(peer);
-            if edit.after.is_none()
-                && let Some(before) = edit.before.as_ref().filter(|entry| entry.preaccepted())
-            {
-                plan.wake.extend(
-                    before
-                        .transaction
-                        .output_pts_iter()
-                        .map(|point| DependencyKey::Cell(compact_packed(&point))),
-                );
-            }
-            for peer in old_peer.into_iter().chain(new_peer) {
-                peer_footprint.insert(self.route(&peer), false);
-                if old_peer != new_peer {
-                    peer_delta
-                        .entry(peer)
-                        .or_insert_with(|| Vec::with_capacity(1))
-                        .push((
-                            hash.clone(),
-                            (old_peer == Some(peer), new_peer == Some(peer)),
-                        ));
-                }
-            }
-            for entry in edit
-                .before
-                .iter()
-                .chain(&edit.after)
-                .filter(|entry| entry.accepted().is_some())
-            {
-                for point in entry
-                    .transaction
-                    .output_pts_iter()
-                    .chain(entry.transaction.input_pts_iter())
-                {
-                    plan.wake
-                        .insert(DependencyKey::Cell(compact_packed(&point)));
-                }
-            }
-        }
-        // Route edits once before locking; both guarded passes retain the original
-        // shard-major, full-hash order using this bounded borrowed scratch.
-        owner_edits.sort_unstable_by_key(|(index, hash, _)| (*index, *hash));
-        if plan.clear_all || plan.snapshot.is_some() {
-            // The lifecycle writer already excludes ordinary commits. Refresh
-            // proposal counts for unchanged owners in the same snapshot cut.
-            for index in 0..SHARDS {
-                owner_footprint.insert(index, true);
-            }
-        }
-        for hash in plan.reads.owners.keys() {
-            owner_footprint.insert(self.owner_shard(hash), false);
-        }
-        if plan.reads.all.is_some() || plan.reads.accepted.is_some() {
-            for index in 0..SHARDS {
-                owner_footprint.insert(index, false);
-            }
-        }
-        for point in plan.reads.spenders.keys() {
-            dependency_footprint.insert(
-                self.route(&RelationKey::Dependency(DependencyKey::Cell(point.clone()))),
-                false,
-            );
-        }
-        for key in plan.reads.relations.keys() {
-            dependency_footprint.insert(self.route(key), true);
-        }
-        for peer in plan.reads.peers.keys() {
-            peer_footprint.insert(self.route(peer), true);
-        }
-        if let Some((peer, _)) = plan.peer_access {
-            peer_footprint.insert(self.route(&peer), false);
-        }
-        if let Some((peer, _)) = plan.ban {
-            peer_footprint.insert(self.route(&peer), true);
-        }
-        for key in &plan.wake {
-            dependency_footprint.insert(self.route(&RelationKey::Dependency(key.clone())), true);
-        }
-        if let Some(page) = &plan.wake_advance {
-            dependency_footprint
-                .insert(self.route(&RelationKey::Dependency(page.key.clone())), true);
-        }
-        let lifecycle_write = plan.snapshot.is_some();
-        let work_changed = lifecycle_write
-            || plan.edits.values().any(|edit| {
-                edit.after
-                    .as_ref()
-                    .is_some_and(|entry| matches!(entry.phase, Phase::Resolve | Phase::Verify(_)))
-            });
-        let template_changed = lifecycle_write
-            || plan.edits.values().any(|edit| {
-                edit.before
-                    .as_ref()
-                    .is_some_and(|entry| entry.accepted().is_some())
-                    || edit
-                        .after
-                        .as_ref()
-                        .is_some_and(|entry| entry.accepted().is_some())
-            });
-        if !plan.committed.is_empty() && !lifecycle_write {
-            return Err(Error::Fault("committed hash source"));
-        }
-        #[cfg(test)]
-        let observer = self.commit_observer.lock().clone();
-        #[cfg(test)]
-        if let Some(observer) = &observer {
-            observer(plan, false);
-        }
-        #[cfg(feature = "profiling")]
-        let acquire_span =
-            tracing::trace_span!(target: "ckb_tx_pool_profile", "tx_pool.authority.acquire")
-                .entered();
-        let mut view = if lifecycle_write {
-            Guard::Write(self.view.write())
-        } else {
-            Guard::Read(self.view.read())
-        };
-        let peer_guards = acquire(&self.peer_gates, &peer_footprint);
-        let dependency_guards = acquire(&self.dependency_gates, &dependency_footprint);
-        let mut owners = acquire(&self.shards, &owner_footprint);
-        #[cfg(feature = "profiling")]
-        drop(acquire_span);
-        if view.get().revision != plan.view {
-            return Err(Error::Stale);
-        }
-        if lifecycle_write && view.get().revision == u64::MAX {
-            return Err(Error::Fault("view counter"));
-        }
-        if self.chain_pending.load(Ordering::Acquire) && !lifecycle_write && !plan.edits.is_empty()
-        {
-            return Err(Error::Full(FullReason::ChainTransition));
-        }
-        if plan.clear_all
-            && (!lifecycle_write
-                || plan.reads.all.is_none()
-                || plan.edits.values().any(|edit| edit.after.is_some())
-                || plan.edits.len()
-                    != owners
-                        .iter()
-                        .map(|(_, guard)| guard.get().owners.len())
-                        .sum::<usize>())
-        {
-            return Err(Error::Fault("incomplete generation replacement"));
-        }
-        self.validate_reads(&plan.reads, &owners)?;
-        #[cfg(test)]
-        if let Some(observer) = &observer {
-            observer(plan, true);
-        }
-        let now = Instant::now();
-        if let Some((peer, expected)) = plan.peer_access
-            && self
-                .bans
-                .lock()
-                .get(&peer)
-                .is_some_and(|until| *until > now)
-                != expected
-        {
-            return Err(Error::Stale);
-        }
-        for edit in plan.edits.values() {
-            if let Some(entry) = edit.after.as_deref()
-                && let Some(peer) = peer(entry)
-                && self
-                    .bans
-                    .lock()
-                    .get(&peer)
-                    .is_some_and(|deadline| *deadline > now)
-            {
-                return Err(Error::Full("peer is banned".into()));
-            }
-        }
-        if let Some(page) = &plan.wake_advance {
-            let row = self
-                .relation(&RelationKey::Dependency(page.key.clone()))
-                .ok_or(Error::Stale)?;
-            if !page.row.ptr_eq(&Arc::downgrade(&row))
-                || row
-                    .lock()
-                    .wake
-                    .as_ref()
-                    .is_none_or(|wake| wake.pass != page.pass)
-            {
-                return Err(Error::Stale);
-            }
-        }
-        let mut remaining_edits = owner_edits.as_slice();
-        for (index, guard) in &owners {
-            // Every edited shard has a write guard; read guards have no edits.
-            let Guard::Write(shard) = guard else {
-                continue;
-            };
-            let (edits, remaining) = remaining_edits.split_at(
-                remaining_edits.partition_point(|(edit_index, _, _)| edit_index == index),
-            );
-            remaining_edits = remaining;
-            let changes = edits.len() as u64;
-            let accepted_changes = edits
-                .iter()
-                .filter(|(_, _, edit)| {
-                    edit.before
-                        .as_ref()
-                        .is_some_and(|entry| entry.accepted().is_some())
-                        || edit
-                            .after
-                            .as_ref()
-                            .is_some_and(|entry| entry.accepted().is_some())
-                })
-                .count() as u64;
-            shard
-                .revision
-                .checked_add(changes)
-                .ok_or(Error::Fault("owner revision"))?;
-            shard
-                .accepted_revision
-                .checked_add(accepted_changes)
-                .ok_or(Error::Fault("accepted revision"))?;
-            // Full-hash order keeps equal proposal prefixes adjacent and in
-            // this same shard. Compare prospective owners, skipping removals.
-            let mut previous_proposal = None;
-            for (_, hash, edit) in edits.iter().copied() {
-                if edit.after.is_some() {
-                    let proposal = proposal_key(hash);
-                    if let Some(existing) = shard.proposals.get(proposal)
-                        && existing != hash
-                        && plan
-                            .edits
-                            .get(existing)
-                            .is_none_or(|edit| edit.after.is_some())
-                    {
-                        return Err(Error::Full("proposal short-ID collision".into()));
-                    }
-                    if previous_proposal == Some(proposal) {
-                        return Err(Error::Full("proposal short-ID collision".into()));
-                    }
-                    previous_proposal = Some(proposal);
-                }
-            }
-        }
-        // Projection contradictions are discovered before mutation. Point
-        // readers may update different members under compatible dependency gates.
-        for (key, changes) in &delta {
-            let row = self.relation(key);
-            let row = row.as_ref().map(|row| row.lock());
-            let mut spender = row.as_ref().and_then(|row| row.spender.clone());
-            for (hash, (old, _)) in changes {
-                if row.as_ref().map_or(0, |row| row.roles(hash)) != *old {
-                    return Err(Error::Fault("relation projection"));
-                }
-                if old & INPUT != 0 && spender.as_ref() == Some(hash) {
-                    spender = None;
-                }
-            }
-            for (hash, (_, new)) in changes {
-                if new & INPUT != 0 {
-                    if spender.as_ref().is_some_and(|old| old != hash) {
-                        return Err(Error::Fault("multiple accepted spenders"));
-                    }
-                    spender = Some(hash.clone());
-                }
-            }
-        }
-        for key in &plan.wake {
-            if let Some(row) = self.relation(&RelationKey::Dependency(key.clone())) {
-                row.lock()
-                    .next_pass
-                    .checked_add(1)
-                    .ok_or(Error::Fault("wake pass"))?;
-            }
-        }
-        for (peer, changes) in &peer_delta {
-            let row = self.peers.lock().get(peer).cloned();
-            let row = row.as_ref().map(|row| row.lock());
-            for (hash, (old, _)) in changes {
-                if row.as_ref().is_some_and(|row| row.members.contains(hash)) != *old {
-                    return Err(Error::Fault("peer projection"));
-                }
-            }
-        }
-        // Scalar capacity is reserved only after every owner observation is
-        // validated. This reservation commits or drops before these guards open.
-        // A sparse admission may lose room to a disjoint commit; replan its fee
-        // policy against the now-settled population instead of rejecting it.
-        let budget = owner_delta
-            .reserve(&self.budget)
-            .map_err(|error| match error {
-                Error::Full(FullReason::Accepted)
-                    if plan.reads.all.is_none() && plan.reads.accepted.is_none() =>
-                {
-                    Error::Stale
-                }
-                error => error,
-            })?;
-        if plan.dry_run {
-            return Ok(None);
-        }
-        let mut batch = None;
-        commit_infallibly(|| {
-            // All recoverable checks are complete. Collection and version-marker
-            // allocation use Rust's abort-on-OOM platform contract.
-            let mut retired = Vec::with_capacity(plan.edits.len());
-            let mut retired_shards = Vec::new();
-            let mut retired_relations = Vec::new();
-            let mut retired_peers = None;
-            let mut retired_dirty = None;
-            let mut retired_queues = None;
-            let mut retired_snapshot = None;
-            let mut retired_committed = None;
-            if plan.clear_all {
-                retired_committed = Some(std::mem::replace(
-                    &mut *self.committed.lock(),
-                    lru::LruCache::new(100_000),
-                ));
-                retired_shards.reserve_exact(SHARDS);
-                retired_relations.reserve_exact(SHARDS);
-                for (_, guard) in &mut owners {
-                    if let Some(shard) = guard.get_mut() {
-                        retired_shards.push(std::mem::take(shard));
-                    }
-                }
-                for relations in &self.relations {
-                    retired_relations.push(std::mem::take(&mut *relations.lock()));
-                }
-                retired_peers = Some(std::mem::take(&mut *self.peers.lock()));
-                retired_dirty = Some(std::mem::take(&mut *self.dirty.lock()));
-                retired_queues = Some(self.queues.take());
-            } else {
-                let mut remaining_edits = owner_edits.as_slice();
-                for (index, guard) in &mut owners {
-                    let Some(shard) = guard.get_mut() else {
-                        continue;
-                    };
-                    let (edits, remaining) = remaining_edits.split_at(
-                        remaining_edits.partition_point(|(edit_index, _, _)| edit_index == index),
-                    );
-                    remaining_edits = remaining;
-                    for (_, hash, edit) in edits.iter().copied() {
-                        let proposal = proposal_key(hash);
-                        let before_deadline = edit.before.as_deref().and_then(deadline);
-                        let after_deadline = edit.after.as_deref().and_then(deadline);
-                        if let Some(before) = &edit.before {
-                            // Exact owner validation and the population bound
-                            // make these derived phase-count updates infallible.
-                            shard.orphan -= usize::from(matches!(before.phase, Phase::Waiting(_)));
-                            shard.proposed -= usize::from(before.accepted().is_some_and(|value| {
-                                value.status(&view.get().snapshot) == Status::Proposed
-                            }));
-                            // A prior edit in this same plan may already have
-                            // transferred this proposal ID to its new owner.
-                            if edit.after.is_none() && shard.proposals.get(proposal) == Some(hash) {
-                                shard.proposals.remove(proposal);
-                            }
-                            if before_deadline != after_deadline
-                                && let Some(deadline) = before_deadline
-                            {
-                                shard.deadlines.remove(&(deadline, hash.clone()));
-                            }
-                            if let Some(accepted) = before.accepted() {
-                                shard
-                                    .accepted_times
-                                    .remove(&(accepted.timestamp, hash.clone()));
-                            }
-                            if matches!(before.phase, Phase::Resolve | Phase::Verify(_)) {
-                                self.queues.remove(before);
-                            }
-                        }
-                        if let Some(after) = &edit.after {
-                            shard.orphan += usize::from(matches!(after.phase, Phase::Waiting(_)));
-                            shard.proposed += usize::from(after.accepted().is_some_and(|value| {
-                                value.status(&view.get().snapshot) == Status::Proposed
-                            }));
-                            // An owner update keeps its hash and proposal mapping.
-                            if edit.before.is_none() {
-                                shard.proposals.insert(*proposal, hash.clone());
-                            }
-                            if before_deadline != after_deadline
-                                && let Some(deadline) = after_deadline
-                            {
-                                shard.deadlines.insert((deadline, hash.clone()));
-                            }
-                            if let Some(accepted) = after.accepted() {
-                                shard
-                                    .accepted_times
-                                    .insert((accepted.timestamp, hash.clone()));
-                            }
-                            retired.extend(shard.owners.insert(hash.clone(), Arc::clone(after)));
-                            if matches!(after.phase, Phase::Resolve | Phase::Verify(_)) {
-                                self.queues.insert(after);
-                            }
-                        } else {
-                            retired.extend(shard.owners.remove(hash));
-                        }
-                        shard.revision += 1;
-                        if edit
-                            .before
-                            .as_ref()
-                            .is_some_and(|entry| entry.accepted().is_some())
-                            || edit
-                                .after
-                                .as_ref()
-                                .is_some_and(|entry| entry.accepted().is_some())
-                        {
-                            shard.accepted_revision += 1;
-                        }
-                    }
-                }
-            }
-            for (key, changes) in delta {
-                self.apply_relation(key, changes, &plan.edits, &plan.wake);
-            }
-            for (peer, changes) in peer_delta {
-                let mut peers = self.peers.lock();
-                let row = peers
-                    .entry(peer)
-                    .or_insert_with(|| Arc::new(Mutex::new(Peer::default())));
-                let mut row = row.lock();
-                for (hash, (_, new)) in changes {
-                    if new {
-                        row.members.insert(hash);
-                    } else {
-                        row.members.remove(&hash);
-                    }
-                }
-                row.version = Arc::new(());
-                let empty = row.members.is_empty();
-                drop(row);
-                if empty {
-                    peers.remove(&peer);
-                }
-            }
-            if let Some((peer, deadline)) = plan.ban {
-                let mut bans = self.bans.lock();
-                bans.retain(|_, deadline| *deadline > now);
-                if bans.len() >= crate::constants::PEER_BAN_FENCE_CAPACITY
-                    && !bans.contains_key(&peer)
-                    && let Some(oldest) = bans
-                        .iter()
-                        .min_by_key(|(_, deadline)| **deadline)
-                        .map(|(peer, _)| *peer)
-                {
-                    bans.remove(&oldest);
-                }
-                bans.entry(peer)
-                    .and_modify(|current| *current = (*current).max(deadline))
-                    .or_insert(deadline);
-            }
-            for key in &plan.wake {
-                self.start_wake(key);
-            }
-            if let Some(page) = &plan.wake_advance {
-                self.advance_wake(page);
-            }
-            if let Some(snapshot) = &plan.snapshot {
-                for (_, guard) in &mut owners {
-                    if let Some(shard) = guard.get_mut() {
-                        shard.proposed = shard
-                            .owners
-                            .values()
-                            .filter(|entry| {
-                                entry
-                                    .accepted()
-                                    .is_some_and(|value| value.status(snapshot) == Status::Proposed)
-                            })
-                            .count();
-                    }
-                }
-            }
-            if let Some(view) = view.get_mut() {
-                if let Some(snapshot) = plan.snapshot.take() {
-                    retired_snapshot = Some(std::mem::replace(&mut view.snapshot, snapshot));
-                }
-                view.revision += 1;
-            }
-            if !plan.committed.is_empty() {
-                let mut cache = self.committed.lock();
-                for (proposal, hash) in std::mem::take(&mut plan.committed) {
-                    cache.put(proposal, hash);
-                }
-            }
-            batch = notice.take().map(notice::Reservation::append);
-            let capacity_returned = budget.commit();
-            drop(owners);
-            drop(dependency_guards);
-            drop(peer_guards);
-            drop(view);
-            if let Some(batch) = &batch {
-                batch.activate(&self.outbox);
-            }
-            if capacity_returned {
-                self.budget.changed.notify_waiters();
-            }
-            if template_changed {
-                self.template_changed.notify_waiters();
-            }
-            // Queue-only transitions wake workers below. Maintenance needs
-            // dependency, capacity or lifecycle progress instead.
-            if lifecycle_write || !plan.wake.is_empty() || capacity_returned {
-                self.changed.notify_waiters();
-            }
-            if work_changed {
-                self.work.notify_waiters();
-            }
-            drop((
-                retired,
-                retired_shards,
-                retired_relations,
-                retired_peers,
-                retired_dirty,
-                retired_queues,
-                retired_snapshot,
-                retired_committed,
-            ));
-        });
-        if self.is_faulted() {
-            return Err(Error::Fault("committed projection"));
-        }
-        Ok(batch)
-    }
-
     #[expect(
         clippy::indexing_slicing,
         reason = "Guard indices originate from the same SHARDS-sized routing as revision arrays."
@@ -1981,161 +985,5 @@ impl Store {
             }
         }
         Ok(())
-    }
-    #[expect(
-        clippy::indexing_slicing,
-        reason = "Keyed routing modulo SHARDS indexes fixed arrays of exactly SHARDS buckets."
-    )]
-    fn apply_relation(
-        &self,
-        key: RelationKey,
-        changes: MemberChanges<u8>,
-        edits: &BTreeMap<Byte32, Edit>,
-        wake: &BTreeSet<DependencyKey>,
-    ) {
-        let mut collection = self.relations[self.route(&key)].lock();
-        let row = collection
-            .entry(key.clone())
-            .or_insert_with(|| Arc::new(Mutex::new(Relation::default())));
-        let mut row = row.lock();
-        for (hash, (old, _)) in &changes {
-            if old & INPUT != 0 && row.spender.as_ref() == Some(hash) {
-                row.spender = None;
-            }
-        }
-        // Use the final spender, independent of the order of owner hashes.
-        // A newly blocked history cannot recover during its creation event.
-        let spent = row.spender.is_some() || changes.iter().any(|(_, (_, new))| new & INPUT != 0);
-        let accepted_changed = changes
-            .iter()
-            .any(|(_, (old, new))| old & ACCEPTED_ROLES != new & ACCEPTED_ROLES);
-        for (hash, (_, new)) in changes {
-            let other_roles = new & !INPUT;
-            if other_roles == 0 {
-                row.members.remove(&hash);
-            } else {
-                let deferred = matches!(&key, RelationKey::Dependency(key) if wake.contains(key))
-                    && edits
-                        .get(&hash)
-                        .and_then(|edit| edit.after.as_ref())
-                        .is_some_and(|entry| {
-                            matches!(
-                                entry.phase,
-                                Phase::Replaced {
-                                    require_all,
-                                    ..
-                                } if !require_all || spent
-                            )
-                        });
-                let wait_after_pass = match row.next_pass.checked_add(u64::from(deferred)) {
-                    Some(pass) => pass,
-                    None => {
-                        self.faulted.store(true, Ordering::Release);
-                        row.next_pass
-                    }
-                };
-                row.members.insert(
-                    hash.clone(),
-                    RelationMember {
-                        roles: other_roles,
-                        wait_after_pass,
-                    },
-                );
-            }
-            if new & INPUT != 0 {
-                row.spender = Some(hash);
-            }
-        }
-        if accepted_changed {
-            row.accepted_version = Arc::new(());
-        }
-        if row.members.is_empty() {
-            // Release the empty BTreeMap root leaf for a spender-only row.
-            row.members.clear();
-            if row.spender.is_none()
-                && row.wake.take().is_some()
-                && let RelationKey::Dependency(key) = &key
-            {
-                self.dirty.lock().remove(key);
-            }
-        }
-        let empty = row.is_empty();
-        drop(row);
-        if empty {
-            collection.remove(&key);
-        }
-    }
-    #[expect(
-        clippy::indexing_slicing,
-        reason = "Keyed routing modulo SHARDS indexes fixed arrays of exactly SHARDS buckets."
-    )]
-    fn start_wake(&self, key: &DependencyKey) {
-        let relation_key = RelationKey::Dependency(key.clone());
-        let collection = self.relations[self.route(&relation_key)].lock();
-        let Some(row) = collection.get(&relation_key) else {
-            return;
-        };
-        let mut row = row.lock();
-        let was_queued = row.wake.is_some();
-        let mut waiters = row
-            .members
-            .values()
-            .filter(|member| member.roles & WAIT != 0);
-        let Some(first) = waiters.next() else {
-            return;
-        };
-        if let Some(pass) = row.next_pass.checked_add(1) {
-            let eligible =
-                pass > first.wait_after_pass || waiters.any(|member| pass > member.wait_after_pass);
-            // Consume even a deferred event so the next event is eligible.
-            row.next_pass = pass;
-            if !eligible {
-                return;
-            }
-            row.wake = Some(Wake { pass, after: None });
-        } else {
-            self.faulted.store(true, Ordering::Release);
-            return;
-        }
-        // Wake and dirty membership change under this same row lock. A newer
-        // pass keeps the existing key and does not need another dirty lookup.
-        if !was_queued {
-            self.dirty.lock().insert(compact_dependency(key));
-        }
-    }
-    #[expect(
-        clippy::indexing_slicing,
-        reason = "Keyed routing modulo SHARDS indexes fixed arrays of exactly SHARDS buckets."
-    )]
-    fn advance_wake(&self, page: &WakePage) {
-        let key = RelationKey::Dependency(page.key.clone());
-        let mut collection = self.relations[self.route(&key)].lock();
-        let Some(row) = collection.get(&key) else {
-            // Preflight validated the row; this commit retired its last member.
-            return;
-        };
-        let mut row = row.lock();
-        // This same commit may start a newer pass; an old cursor cannot erase it.
-        if row.wake.as_ref().is_none_or(|wake| wake.pass != page.pass) {
-            return;
-        }
-        let more = row
-            .members
-            .range((page.after.as_ref().map_or(Unbounded, Excluded), Unbounded))
-            .any(|(_, flags)| flags.roles & WAIT != 0 && page.pass > flags.wait_after_pass);
-        if more {
-            row.wake = Some(Wake {
-                pass: page.pass,
-                after: page.after.clone(),
-            });
-        } else {
-            row.wake = None;
-            self.dirty.lock().remove(&page.key);
-        }
-        let empty = row.is_empty();
-        drop(row);
-        if empty {
-            collection.remove(&key);
-        }
     }
 }
